requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymupdf==1.23.8
google-genai==0.2.2
python-dotenv==1.0.0

